ANW-148: Remove OCLC search from LCNAF plugin
