+++ frontend/app/views/shared/_templates.html.erb
@@ -19,7 +19,7 @@
     ${message}
   </div>
   <div class="modal-footer">
-    <button class="btn" data-dismiss="modal"><%= I18n.t "actions.continue" %></button>
+    <button class="btn btn-default" data-dismiss="modal"><%= I18n.t "actions.continue" %></button>
   </div>
 --></div>
