+++ plugins/lcnaf/frontend/models/opensearcher.rb
@@ -37,7 +37,6 @@ class OpenSearcher
     lccns.each do |lccn|
       lccn.sub!( 'info:lc/authorities/subjects/', '')
       uri = URI("#{@scheme}/#{lccn}.marcxml.xml")
-      p uri
 
       HTTPRequest.new.get(uri) do |response|
         if response.code != '200'
@@ -72,7 +71,6 @@ class OpenSearcher
                                   'start' => start_record)
 
     uri.query = URI.encode_www_form(params)
-    p uri
     results = HTTPRequest.new.get(uri) do |response|
       if response.code != '200'
         raise OpenSearchException.new("Error during OpenSearch search: #{response.body}")
