+++ plugins/lcnaf/frontend/spec/spec_helper.rb
@@ -1,0 +1,9 @@
+require 'rspec'
+require 'nokogiri'
+
+
+def parse(xml_string)
+  doc = Nokogiri::XML.parse(xml_string)
+  doc.remove_namespaces!
+  doc
+end
