+++ plugins/lcnaf/frontend/views/lcnaf/index.html.erb
@@ -21,12 +21,6 @@
            LCSH - http://id.loc.gov/authorities/subjects
           </label>
       </div>
-      <div class="radio">
-        <label for="lcnaf_service_oclc">
-          <input type="radio" name="lcnaf_service" value="oclc" id="lcnaf_service_oclc">
-          OCLC - http://alcme.oclc.org/srw/search/lcnaf
-        </label>
-      </div>
     </div>
 
     <div class='control-group form-group required'>
