--- plugins/lcnaf/spec/mocks/n82024365.marcxml.xml
@@ -1,68 +1,0 @@
-<marcxml:record xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:marcxml="http://www.loc.gov/MARC21/slim" xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:madsrdf="http://www.loc.gov/mads/rdf/v1#" xmlns:ri="http://id.loc.gov/ontologies/RecordInfo#" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:mets="http://www.loc.gov/METS/">
-	    <marcxml:leader>01470cz  a2200301n  4500</marcxml:leader>
-	    <marcxml:controlfield tag="001">n82024365</marcxml:controlfield>
-	    <marcxml:controlfield tag="003">DLC</marcxml:controlfield>
-	    <marcxml:controlfield tag="005">20080411152524.0</marcxml:controlfield>
-	    <marcxml:controlfield tag="008">820326n| acannaabn          |a aaa      </marcxml:controlfield>
-	    <marcxml:datafield tag="010" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">n  82024365 </marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="035" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">(OCoLC)oca00721546</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="040" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">DLC</marcxml:subfield>
-    <marcxml:subfield code="b">eng</marcxml:subfield>
-    <marcxml:subfield code="c">DLC</marcxml:subfield>
-    <marcxml:subfield code="d">DLC</marcxml:subfield>
-    <marcxml:subfield code="d">OCoLC</marcxml:subfield>
-    <marcxml:subfield code="d">DLC</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="100" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Machado, Franklin de C.</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Vitória de Cerqueira Barreiros Machado, Franklin</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">De Cerqueira Barreiros Machado, Franklin Vitória</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Cerqueira Barreiros Machado, Franklin Vitória de</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Barreiros Machado, Franklin Vitória de Cerqueira</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Maxado, F.</marcxml:subfield>
-    <marcxml:subfield code="q">(Franklin)</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Maxado, Franklin</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Maxado Nordestino, Franklin</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="400" ind1="1" ind2=" ">
-    <marcxml:subfield code="a">Nordestino, Franklin Maxado</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="670" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">His Feira de Santana, 1967?</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="670" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">His O que é literatura de cordel?, 1980:</marcxml:subfield>
-    <marcxml:subfield code="b">t.p. (Franklin Maxado) front flap (Franklin Machado)</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="670" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">nuc87-132066: His Malvinas são malvindas, 1982</marcxml:subfield>
-    <marcxml:subfield code="b">(hdg. on NmU rept.: Machado, Franklin de C.; usage: Franklin Maxado Nordestino)</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="670" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">nuc89-79085: His O jumento que virou gente ... 1979?</marcxml:subfield>
-    <marcxml:subfield code="b">(hdg. on NmU rept.: Machado, Franklin de C.; usage: F. Maxado)</marcxml:subfield>
-  </marcxml:datafield>
-	    <marcxml:datafield tag="670" ind1=" " ind2=" ">
-    <marcxml:subfield code="a">Franklin Maxado, 2007:</marcxml:subfield>
-    <marcxml:subfield code="b">t.p. verso (Braz. CIP hdg.: Machado, Franklin, 1943- ) front flap (Franklin Maxado; Franklin de Cerqueira Machado; Maxado Nordestino) p. 9 (Franklin Vitória de Cerqueira Barreiros Machado; b. Mar. 15, 1943, in Feira de Santana)</marcxml:subfield>
-  </marcxml:datafield>
-	  </marcxml:record>
